Coincidence_DAQ
This code is able to do online coincidence selection on-line. The user can select their desired window size and the threshold. If multiple waves occur within the selection window, the code will calculate the "matrix" which is based on their distance and then select the pair with the largest "matrix".